[Core] Add CMake Project for MediaCreation
As Traitgen is now getting support for C++ generated traits,
Mediacreation wants to provide a package of these generated
traits, easily usable for C++ hosts and resolvers.

This commit adds a cmake workflow to generate a mediacreation
cmake package, which can be included into a build pipeline
via `find_package`, or via directory include as a submodule.

Signed-off-by: Elliot Morris <elliot.morris@foundry.com>